refactor: split out run method in standalone function, spawn consensus api directly
